backend/src/core/serializers.py
from core.models import Conversation, ConversationMessage
from rest_framework import serializers


class ConversationMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ConversationMessage
        fields = (
            "id",
            "type",
            "message",
        )


class ConversationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Conversation
        fields = (
            "id",
            "title",
        )


class ConversationCreateSerializer(serializers.Serializer):
    message = serializers.CharField()
    conversation_id = serializers.IntegerField(required=False)
