backend/pyproject.toml
[tool.poetry]
name = "webscope"
version = "0.1.0"
description = ""
authors = ["andrejbinovsky <binovsky.andrej@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg = { extras = ["binary"], version = "^3.1.19" }
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
openai = "^1.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
